first 98 fib num

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,7 +10,7 @@
 int main(void)
 {
 	int a = 0;
-	long j =1, k = 2;
+	float j = 1, k = 2;
 
 	while (a < 98)
 	{
